server/requirements.txt
aim-cli==2.0.6
aimrecords==0.0.4
alembic==1.1.0
aniso8601==7.0.0
certifi==2019.6.16
cffi==1.13.1
chardet==3.0.4
Click==7.0
cryptography==2.8
dnspython==1.16.0
email-validator==1.0.5
file-read-backwards==2.0.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-Modus==0.0.1
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
tensorboard==1.15.0
psycopg2==2.8.3
pycparser==2.19
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.2
protobuf>=3.11.0
tornado==4.5.3
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.8
urllib3==1.25.3
Werkzeug==0.15.5
WTForms==2.2.1
